+++ CHANGELOG.md
@@ -11,3 +11,9 @@ For any change that affects end users of this package, please add an entry under
 If your change does not need a CHANGELOG entry, add the "skip changelog" label to your PR.
 
 ## Unreleased
+- Add botocore instrumentation extension for Bedrock AgentCore services with span attributes
+  ([#490](https://github.com/aws-observability/aws-otel-python-instrumentation/pull/490))
+- [PATCH] Only decode JSON input buffer in Anthropic Claude streaming
+    ([#497](https://github.com/aws-observability/aws-otel-python-instrumentation/pull/497))
+- Fix timeout handling for exceeded deadline in retry logic in OTLPAwsLogsExporter
+  ([#501](https://github.com/aws-observability/aws-otel-python-instrumentation/pull/501))